latest

+++ softmax.py
@@ -1,5 +1,5 @@
 import numpy as np
 
 def softmax(w):
-    e = np.exp(w).T
+    e = np.exp(w.astype(float)).T
     return (e / np.sum(e, axis = 0)).T
